Separate Sql service specific log out repo
I want to write simulator for the sql service. But I realize that I cannot
simply separate the specific Sql service API and make it an ISqlSimulator.
Because a lot of logic dependent on the API class.

Then I find that in Repository(Yucky) it is acutally do at least 2
different thing: 1. make the sql connection and do make query, read
result.

So I separate them. Since IDesposible is an interesting case, I need to
use more interesting lambda to achieve the separation